Add test adoption sol

+++ test/TestAdoption.sol
@@ -9,31 +9,23 @@ contract TestAdoption {
     Adoption adoption = Adoption(DeployedAddresses.Adoption());
 
     // The id of the pet that will be used for testing
-    uint256 expectedPetId = 8;
+    uint expectedPetId = 8;
 
     // The expected owner of adopted pet is this contract
     address expectedAdopter = address(this);
 
     // Testing the adopt() function
     function testUserCanAdoptPet() public {
-        uint256 returnedId = adoption.adopt(expectedPetId);
+        uint returnedId = adoption.adopt(expectedPetId);
 
-        Assert.equal(
-            returnedId,
-            expectedPetId,
-            "Adoption of the expected pet should match what is returned."
-        );
+        Assert.equal(returnedId, expectedPetId, "Adoption of the expected pet should match what is returned.");
     }
 
     // Testing retrieval of a single pet's owner
     function testGetAdopterAddressByPetId() public {
         address adopter = adoption.adopters(expectedPetId);
 
-        Assert.equal(
-            adopter,
-            expectedAdopter,
-            "Owner of the expected pet should be this contract"
-        );
+        Assert.equal(adopter, expectedAdopter, "Owner of the expected pet should be this contract");
     }
 
     // Testing retrieval of all pet owners
@@ -41,10 +33,6 @@ contract TestAdoption {
         // Store adopters in memory rather than contract's storage
         address[16] memory adopters = adoption.getAdopters();
 
-        Assert.equal(
-            adopters[expectedPetId],
-            expectedAdopter,
-            "Owner of the expected pet should be this contract"
-        );
+        Assert.equal(adopters[expectedPetId], expectedAdopter, "Owner of the expected pet should be this contract");
     }
 }
